+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-04-15
+
+### Changes
+
+- Bump minumtium to v1.1.0
+- Changed the name of the adapter to MinumtiumSimpleJwtAuth
+- Added a configuration class using Pydantic models instead of using a dict
+
 ## [1.0.1] - 2022-04-15
+
 ### Added
+
 - Added setup of default user in case it does not exist.
 
 ## [1.0.0] - 2022-04-08
+
 ### Added
+
 - Initial release